Переименовал все Window в View, раскидал по папкам ViewModels и Views
