+++ OlympiadWpfApp/OlympiadWpfApp/ViewModels/ShowTableViewModels/ShowSportTypeTableViewModel.cs
@@ -4,16 +4,18 @@ using System.Runtime.CompilerServices;
 using System.Windows;
 using OlympiadWpfApp.DataAccess.Contexts;
 using OlympiadWpfApp.DataAccess.Entities;
+using OlympiadWpfApp.ViewModels.ViewRowViewModels;
 using OlympiadWpfApp.Views;
+using OlympiadWpfApp.Views.ViewRowViews;
 
-namespace OlympiadWpfApp.ViewModels;
+namespace OlympiadWpfApp.ViewModels.ShowTableViewModels;
 
-public class ShowOlympiadTableViewModel : ShowTableViewModel, INotifyPropertyChanged
+public class ShowSportTypeTableViewModel : ShowTableViewModel
 {
-    private readonly OlympDbContext _olympDbContext;
-    public ObservableCollection<OlympiadEntity> Entities { get; private set; }
+     private readonly OlympDbContext _olympDbContext;
+    public ObservableCollection<SportTypeEntity> Entities { get; private set; }
 
-    public ShowOlympiadTableViewModel(Window owner, OlympDbContext olympDbContext) : base(owner)
+    public ShowSportTypeTableViewModel(Window owner, OlympDbContext olympDbContext) : base(owner)
     {
         _olympDbContext = olympDbContext;
         GetData();
@@ -24,8 +26,8 @@ public class ShowOlympiadTableViewModel : ShowTableViewModel, INotifyPropertyCha
         var index = SelectedIndex;
         var selectedEntity = Entities[index]; // пытался сделать клон сущности(чтобы в случае отмена откатить изменния), но получал ошибку с отслеживанием изменений, выкрутился костылями
 
-        var window = new ViewOlympiadRowWindow(Owner);
-        var viewModel = new ViewOlympiadRowViewModel(window, selectedEntity);
+        var window = new ViewSportTypeRowView(Owner);
+        var viewModel = new ViewSportTypeRowViewModel(window, selectedEntity);
 
         if (window.ShowDialog() != true)
         {
@@ -33,41 +35,35 @@ public class ShowOlympiadTableViewModel : ShowTableViewModel, INotifyPropertyCha
             return;
         }
 
-        _olympDbContext.Olympiads.Update(selectedEntity);
+        _olympDbContext.SportTypes.Update(selectedEntity);
     }
 
     protected override void ExecuteAdd()
     {
-        var window = new ViewOlympiadRowWindow(Owner);
-        var olympiadEntity = new OlympiadEntity
+        var sportTypeEntity = new SportTypeEntity()
         {
-            Id = _olympDbContext.Olympiads.Any() ? _olympDbContext.Olympiads.OrderBy(x => x.Id).Last().Id + 1 : 1,
+            Id = _olympDbContext.SportTypes.Any() ? _olympDbContext.SportTypes.OrderBy(x => x.Id).Last().Id + 1 : 1,
             Name = "",
-            Year = DateOnly.Parse("01.01.1900"),
-            HostCountry = "",
-            City = "",
-            IsWinter = false,
-            IsDeleted = false,
         };
 
-        var viewModel = new ViewOlympiadRowViewModel(window, olympiadEntity);
+        var window = new ViewSportTypeRowView(Owner);
+        var viewModel = new ViewSportTypeRowViewModel(window, sportTypeEntity);
 
         if (window.ShowDialog() != true) return;
 
-        _olympDbContext.Olympiads.Add(viewModel.Entity);
+        _olympDbContext.SportTypes.Add(viewModel.Entity);
         Entities.Add(viewModel.Entity);
     }
 
     protected override void ExecuteDelete()
     {
-        Entities[SelectedIndex].IsDeleted = true; // Надеюсь правильно понял soft-delete механизм
+        _olympDbContext.SportTypes.Remove(Entities[SelectedIndex]);
         Entities.Remove(Entities[SelectedIndex]);
     }
 
     protected override void GetData()
     {
-        Entities = new ObservableCollection<OlympiadEntity>(_olympDbContext.Olympiads
-            .Where(x => !x.IsDeleted)
+        Entities = new ObservableCollection<SportTypeEntity>(_olympDbContext.SportTypes
             .OrderBy(x => x.Id)
             .ToList());
     }
