+++ OlympiadWpfApp/OlympiadWpfApp/ViewModels/ViewRowViewModels/ViewSportTypeRowViewModel.cs
@@ -2,7 +2,7 @@ using System.Windows;
 using OlympiadWpfApp.Commands;
 using OlympiadWpfApp.DataAccess.Entities;
 
-namespace OlympiadWpfApp.ViewModels;
+namespace OlympiadWpfApp.ViewModels.ViewRowViewModels;
 
 public class ViewSportTypeRowViewModel
 {
