+++ OlympiadWpfApp/OlympiadWpfApp/Views/MainView.xaml.cs
@@ -5,9 +5,9 @@ namespace OlympiadWpfApp.Views;
 /// <summary>
 /// Interaction logic for MainWindow.xaml
 /// </summary>
-public partial class MainWindow : Window
+public partial class MainView : Window
 {
-    public MainWindow()
+    public MainView()
     {
         InitializeComponent();
     }
