--- OlympiadWpfApp/OlympiadWpfApp/Views/ShowOlympiadTableWindow.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows;
-
-namespace OlympiadWpfApp.Views;
-
-public partial class ShowOlympiadTableWindow : Window
-{
-    public ShowOlympiadTableWindow(Window owner)
-    {
-        Owner = owner;
-        InitializeComponent();
-    }
-}
