--- OlympiadWpfApp/OlympiadWpfApp/Views/ShowParticipantTableWindow.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows;
-
-namespace OlympiadWpfApp.Views;
-
-public partial class ShowParticipantTableWindow : Window
-{
-    public ShowParticipantTableWindow(Window owner)
-    {
-        Owner = owner;
-        InitializeComponent();
-    }
-}
