--- OlympiadWpfApp/OlympiadWpfApp/Views/ShowSportTypeTableWindow.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows;
-
-namespace OlympiadWpfApp.Views;
-
-public partial class ShowSportTypeTableWindow : Window
-{
-    public ShowSportTypeTableWindow(Window owner)
-    {
-        Owner = owner;
-        InitializeComponent();
-    }
-}
