+++ OlympiadWpfApp/OlympiadWpfApp/Views/ShowTableView/ShowOlympiadTableView.xaml
@@ -1,14 +1,14 @@
-﻿<Window x:Class="OlympiadWpfApp.Views.ShowParticipantTableWindow"
+﻿<Window x:Class="OlympiadWpfApp.Views.ShowTableView.ShowOlympiadTableView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:viewmodels="clr-namespace:OlympiadWpfApp.ViewModels"
         xmlns:entities="clr-namespace:OlympiadWpfApp.DataAccess.Entities;assembly=OlympiadWpfApp.DataAccess"
         xmlns:converters="clr-namespace:OlympiadWpfApp.Converters"
+        xmlns:showTableViewModels="clr-namespace:OlympiadWpfApp.ViewModels.ShowTableViewModels"
         mc:Ignorable="d"
         Title="Таблица олимпиад" Height="450" Width="600"
-        d:DataContext="{d:DesignInstance viewmodels:ShowParticipantTableViewModel}"
+        d:DataContext="{d:DesignInstance showTableViewModels:ShowOlympiadTableViewModel}"
         WindowStartupLocation="CenterOwner"
         Language="ru-RU">
     <Window.Resources>
@@ -25,58 +25,56 @@
             <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
         </Grid.RowDefinitions>
-        <DataGrid Grid.Row="0" ItemsSource="{Binding Entities, UpdateSourceTrigger=PropertyChanged}"
-                  SelectedIndex="{Binding SelectedIndex}" Margin="10"
+        <DataGrid Grid.Row="0" ItemsSource="{Binding Entities, UpdateSourceTrigger=PropertyChanged}" SelectedIndex="{Binding SelectedIndex}" Margin="10"
                   IsReadOnly="True" AutoGenerateColumns="False">
             <DataGrid.Columns>
                 <DataGridTemplateColumn Header="Id" Width="Auto">
                     <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate DataType="entities:ParticipantEntity">
+                        <DataTemplate DataType="entities:OlympiadEntity">
                             <TextBlock Text="{Binding Id}" />
                         </DataTemplate>
                     </DataGridTemplateColumn.CellTemplate>
                 </DataGridTemplateColumn>
-
-                <DataGridTemplateColumn Header="Фамилия" Width="*">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate DataType="entities:ParticipantEntity">
-                            <TextBlock Text="{Binding Surname}" />
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
                 
-                <DataGridTemplateColumn Header="Имя" Width="*">
+                <DataGridTemplateColumn Header="Название" Width="*">
                     <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate DataType="entities:ParticipantEntity">
+                        <DataTemplate DataType="entities:OlympiadEntity">
                             <TextBlock Text="{Binding Name}" />
                         </DataTemplate>
                     </DataGridTemplateColumn.CellTemplate>
                 </DataGridTemplateColumn>
                 
-                <DataGridTemplateColumn Header="Отчество" Width="*">
+                <DataGridTemplateColumn Header="Дата проведения" Width="Auto">
                     <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate DataType="entities:ParticipantEntity">
-                            <TextBlock Text="{Binding Patronymic}" />
+                        <DataTemplate DataType="entities:OlympiadEntity">
+                            <TextBlock Text="{Binding Year, StringFormat='{}{0:dd.MM.yyyy}'}" />
                         </DataTemplate>
                     </DataGridTemplateColumn.CellTemplate>
                 </DataGridTemplateColumn>
-
-                <DataGridTemplateColumn Header="Дата рождения" Width="Auto">
+                
+                <DataGridTemplateColumn Header="Страна-хозяйка" Width="*">
                     <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate DataType="entities:ParticipantEntity">
-                            <TextBlock Text="{Binding Birthdate, StringFormat='{}{0:dd.MM.yyyy}'}" />
+                        <DataTemplate DataType="entities:OlympiadEntity">
+                            <TextBlock Text="{Binding HostCountry}" />
                         </DataTemplate>
                     </DataGridTemplateColumn.CellTemplate>
                 </DataGridTemplateColumn>
-
-                <DataGridTemplateColumn Header="Страна" Width="*">
+                
+                <DataGridTemplateColumn Header="Город" Width="*">
                     <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate DataType="entities:ParticipantEntity">
-                            <TextBlock Text="{Binding Country}" />
+                        <DataTemplate DataType="entities:OlympiadEntity">
+                            <TextBlock Text="{Binding City}" />
                         </DataTemplate>
                     </DataGridTemplateColumn.CellTemplate>
                 </DataGridTemplateColumn>
-                <!-- Не разобрался, как работать с фото( -->
+                
+                <DataGridTemplateColumn Header="Сезон олимпиады" Width="*">
+                    <DataGridTemplateColumn.CellTemplate>
+                        <DataTemplate DataType="entities:OlympiadEntity">
+                            <TextBlock Text="{Binding IsWinter, Converter={StaticResource SeasonConverter}}" />
+                        </DataTemplate>
+                    </DataGridTemplateColumn.CellTemplate>
+                </DataGridTemplateColumn>
             </DataGrid.Columns>
         </DataGrid>
         <StackPanel Grid.Row="1" Orientation="Horizontal" HorizontalAlignment="Right">
