+++ OlympiadWpfApp/OlympiadWpfApp/Views/ShowTableView/ShowOlympiadTableView.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace OlympiadWpfApp.Views.ShowTableView;
+
+public partial class ShowOlympiadTableView : Window
+{
+    public ShowOlympiadTableView(Window owner)
+    {
+        Owner = owner;
+        InitializeComponent();
+    }
+}
