+++ OlympiadWpfApp/OlympiadWpfApp/Views/ShowTableView/ShowParticipantTableView.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace OlympiadWpfApp.Views.ShowTableView;
+
+public partial class ShowParticipantTableView : Window
+{
+    public ShowParticipantTableView(Window owner)
+    {
+        Owner = owner;
+        InitializeComponent();
+    }
+}
