+++ OlympiadWpfApp/OlympiadWpfApp/Views/ShowTableView/ShowSportTypeTableView.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace OlympiadWpfApp.Views.ShowTableView;
+
+public partial class ShowSportTypeTableView : Window
+{
+    public ShowSportTypeTableView(Window owner)
+    {
+        Owner = owner;
+        InitializeComponent();
+    }
+}
