--- OlympiadWpfApp/OlympiadWpfApp/Views/ViewOlympiadRowWindow.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows;
-
-namespace OlympiadWpfApp.Views;
-
-public partial class ViewOlympiadRowWindow : Window
-{
-    public ViewOlympiadRowWindow(Window owner)
-    {
-        Owner = owner;
-        InitializeComponent();
-    }
-}
