--- OlympiadWpfApp/OlympiadWpfApp/Views/ViewParticipantRowWindow.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows;
-
-namespace OlympiadWpfApp.Views;
-
-public partial class ViewParticipantRowWindow : Window
-{
-    public ViewParticipantRowWindow(Window owner)
-    {
-        Owner = owner;
-        InitializeComponent();
-    }
-}
