+++ OlympiadWpfApp/OlympiadWpfApp/Views/ViewRowViews/ViewOlympiadRowView.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace OlympiadWpfApp.Views.ViewRowViews;
+
+public partial class ViewOlympiadRowView : Window
+{
+    public ViewOlympiadRowView(Window owner)
+    {
+        Owner = owner;
+        InitializeComponent();
+    }
+}
