+++ OlympiadWpfApp/OlympiadWpfApp/Views/ViewRowViews/ViewParticipantRowView.xaml
@@ -1,13 +1,13 @@
-﻿<Window x:Class="OlympiadWpfApp.Views.ViewParticipantRowWindow"
+﻿<Window x:Class="OlympiadWpfApp.Views.ViewRowViews.ViewParticipantRowView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:viewmodels="clr-namespace:OlympiadWpfApp.ViewModels"
         xmlns:converters="clr-namespace:OlympiadWpfApp.Converters"
+        xmlns:viewRowViewModels="clr-namespace:OlympiadWpfApp.ViewModels.ViewRowViewModels"
         mc:Ignorable="d"
         Title="Олимпиада" Height="450" Width="450" SizeToContent="Height"
-        d:DataContext="{d:DesignInstance viewmodels:ViewParticipantRowViewModel}"
+        d:DataContext="{d:DesignInstance viewRowViewModels:ViewParticipantRowViewModel}"
         WindowStartupLocation="CenterOwner"
         Language="ru-RU">
     <Window.Resources>
