+++ OlympiadWpfApp/OlympiadWpfApp/Views/ViewRowViews/ViewParticipantRowView.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace OlympiadWpfApp.Views.ViewRowViews;
+
+public partial class ViewParticipantRowView : Window
+{
+    public ViewParticipantRowView(Window owner)
+    {
+        Owner = owner;
+        InitializeComponent();
+    }
+}
