+++ OlympiadWpfApp/OlympiadWpfApp/Views/ViewRowViews/ViewSportTypeRowView.xaml
@@ -1,13 +1,13 @@
-﻿<Window x:Class="OlympiadWpfApp.Views.ViewSportTypeRowWindow"
+﻿<Window x:Class="OlympiadWpfApp.Views.ViewRowViews.ViewSportTypeRowView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:viewmodels="clr-namespace:OlympiadWpfApp.ViewModels"
+        xmlns:viewRowViewModels="clr-namespace:OlympiadWpfApp.ViewModels.ViewRowViewModels"
         mc:Ignorable="d"
 
         Title="Вид спорта" Height="100" Width="450" SizeToContent="Height"
-        d:DataContext="{d:DesignInstance viewmodels:ViewSportTypeRowViewModel}"
+        d:DataContext="{d:DesignInstance viewRowViewModels:ViewSportTypeRowViewModel}"
         WindowStartupLocation="CenterOwner"
         Language="ru-RU">
     <Window.Resources>
