+++ OlympiadWpfApp/OlympiadWpfApp/Views/ViewRowViews/ViewSportTypeRowView.xaml.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows;
+
+namespace OlympiadWpfApp.Views.ViewRowViews;
+
+public partial class ViewSportTypeRowView : Window
+{
+    public ViewSportTypeRowView(Window owner)
+    {
+        Owner = owner;
+        InitializeComponent();
+    }
+}
