--- OlympiadWpfApp/OlympiadWpfApp/Views/ViewSportTypeRowWindow.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows;
-
-namespace OlympiadWpfApp.Views;
-
-public partial class ViewSportTypeRowWindow : Window
-{
-    public ViewSportTypeRowWindow(Window owner)
-    {
-        Owner = owner;
-        InitializeComponent();
-    }
-}
